pyproject.toml
[project]
name = "pjn-nlp"
version = "1.0.0"
description = "NLP Application for Polish Language Analysis"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pandas",
    "streamlit",
    "pyyaml",
    "spacy",
    "networkx",
    "matplotlib",
    "numpy",
    "beautifulsoup4",
    "requests",
    "pip>=25.3",
    "morfeusz2>=1.99.12",
    'regex',
    'spacy-lookups-data'
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/nlp_app", "src/syntax_generator"]

[tool.uv]
dev-dependencies = ["pytest"]
